initial implementation of OptionComboBox
OptionComboBox is a helper class to aid in creation of a simple ComboBox
consisting only of text, much like Gtk::ComboBoxText, but that can also
make items be sensitive / insensitive.
This is for help in porting from Gtk::OptionMenu in gtkmm2.4, a
combo box that uses a GtkMenu as the underlying model, rather than a
GtkTreeModel. GtkMenu items can be insensitive and that was used
extensively in gparted.